Fix stuck in LruCache after updating items
Inside the LruCache, it always create a new link node when doing the []= operator, what if we're doing the "updating" operation?
The `_entries` will replace the new value to the same key entry, but the link list will have two nodes with a same key, once the capacity is exhausted, the application will stuck in removing item as the second removing with the same key changes nothing.