Update pub.dev README to include Build-Filters
This feature was not displayed in the official README featured on
pub.dev. It was hidden in the CHANGELOG. I essentially copied the
CHANGELOG info and moved it to the README.

By including this information on the main README it will allow
other package users to understand how to configure this package
further.