Fix bug with manual deletions of files from build steps with multiple outputs (#3370)
Fixes #3364

When a manual delete happens, we mark just that output as definitelyNeedsUpdate, but when checking if a given build step needs to run, we were only checking the state of the first output, based on an assumption that all outputs got invalidated together. That is true for all cases except this one.

This changes it to check if any of the output nodes definitely need an update, to cover this particular case.