Automated g4 rollback of changelist 480854047.
*** Reason for rollback ***

Breaks lots of tests across G3

*** Original change description ***

Generate method overrides even then source lib is not null-safe

If the class to mock comes from a null-safe library. Otherwise
we don't get nice mocks (as niceness depends on overrides).

We need to be careful to generate a valid pre-null-safety code,
but so far I only found we need to skip generating `required`.

***

PiperOrigin-RevId: 480867585