Build hidden outputs if they're public (#3390)
Some directories, like `lib/`, are special-cased by the build system in the sense that

1. Their assets are readable for non-root packages
2. Hidden assets are generated for them even if they're not part of any build filter or directory.

Some time ago, I've added the `additional_public_assets` configuration to expand this behavior to non-standard directories. As a motivation, consider building a C library with the build system and some builder generating a hidden file in `include/`. Similar to hidden outputs in `lib/`, we definitely want this builder to run for this asset since `include/` has a meaning somewhat comparable to `lib/` in Dart. The `additional_public_assets` build configuration hence makes it easier to integrate non-Dart components in the Dart package ecosystem and build them with Dart's standard build tools.

I am also running into this issue with a static site generator that I'm writing. It generates a (hidden) asset for markdown files in `pages/` containing HTML and a list of target URLs. A post-process builder later generates `.html` assets in `web/` based on these intermediate assets. This builder breaks when a build directory is present, since hidden assets in `pages/` are no longer built (see https://github.com/dart-lang/webdev/issues/1200, https://github.com/dart-lang/webdev/issues/1733).

Having public assets handled in a consistent way would really help with my static site generator. I also believe that it is generally the right thing to do in the implementation as it could avoid confusion for advanced users interested in building not-only-Dart modules or packages.
So, this PR makes the output check consistent with the notion of a public asset used in other parts of the build implementation.