Move SDK summary generation to a separate library (#3431)
* Move SDK summary generation to a separate library

The details of finding or generating the SDK summary are different
internally, so this code gets patched. The patch file ends up needing
frequent maintenance to keep up with unrelated other changes in this
file. Separate the SDK summary related code into it's own library which
will have a patch so that it is more isolated from other changes.

Move `isFlutter` to `sdk_summary.dart` since it is used there, and it is
related to the SDK. Update the test which imports from `src`.

Move `packagePath` to `build_asset_uri_generator` so it can be used in
both `resolver.dart` and `sdk_summary.dart`.

* dartfmt