Fix the build (#3476)
## build_daemon

We are using `unawaited` which wasn't made available until 2.15, but we can't get a valid solve until `2.18` so I am going to require that.

This also regenerates some built_value code so that it won't trigger lints.

## build_modules

Delete the old dart2js  worker code which has been deprecated for a long time. This is technically breaking but in practice should not be, and I haven't marked it as a breaking change. Feel free to push back.