Let users provide dummy values for types
Mockito needs to create dummy values both during stubbing and
while returning default values from nice mocks. Most cases are
covered by the code generation, but there are some tricky cases
where we either don't know the type at generation time or
the required type can't be implemented, so Mockito cannot easily
fake it. In this cases users could help Mockito by supplying it
with the dummy values.

This change adds two new API functions `provideDummy` and
`provideDummyBuilder` to do that.

Based on work in https://github.com/dart-lang/mockito/pull/592.

I haven't yet removed the checks that fail the generation if a
generic return type is found, so this change doesn't yet allow
people to mock classes that have methods returning type
variables. I plan to lift that restriction in the next change.
So the only actually affected case is the return type being a
function returning a type variable, previouly that would
result in `() => null` and now it's `() => dummyValue<T>()` instead.

I also didn't yet change the handling of `Future`: it seems we
rely heavily on the fakability of `Future`: if codegen sees
`Future<T>` return type, it just creates a `FakeFuture` for it.
It couldn't be `awaited`, but works perfectly fine while
stubbing, so I'm not convienced we want to change it to
`Future.value<T>(dummyValue<T>())` which could fail at run-time,
even if not awaited.

PiperOrigin-RevId: 532712274