Don't abort codegen on return type being a type variable
Now we can continue in this case, hoping that the needed value will be
found at run time. And if not, user will get a message, saying for
which type they have to provide a dummy value.

Tests are extended to cover the new behavior, where possible.
I had to delete tests for fallback generators, as there is nothing
much to test. Fallback generators are now not required for return
type being a type variable. And for private return type they can't
be used due to a bug in codegen: we are trying to cast to a private
type, that is not visible in the generated library.

PiperOrigin-RevId: 534088653