Build runner core 7.2.7+1, backport logging fix (#3521)
Fixes https://github.com/dart-lang/build/issues/3519

Currently, users on the previous flutter SDK (3.7, <6 months old), can't get the latest package:build_runner_core but can get the latest package:logging. This causes compile errors due to a member not being implemented.

This backports the fix to build_runner_core 7.2.7+1, which should be resolvable with build_runner version 2.3.3, which was the last version to allow a Dart version <3.0.0 (it supports 2.14 and later).

Once this is approved, I will tag and release, and then resolve merge conflicts with master so there is a changelog entry.