Allow single builder cycles (#3532)
Fixes #3531

When a builder outputs a file with the same extension as a
`required_input` it will be in a cycle with itself. When builders were
ordered with `stronglyConnectedComponents` cycles were detected by
looking for a component (cycle) with more than one builder. With the
move to `topologicalSort` the single builder cycle causes an exception
during the sort. Exclude the builder itself from the list of
dependencies for the graph algorithm call.