Skip delete for SDK summary and deps file (#3573)
Fixes #2995

When the summary file is determined to be invalid, regenerate it and
overwrite it without deleting. This resolves a race condition where
multiple resolvers could try to delete the same file. The invalid
content will not be read, because we will regenerate the file before
moving on to the rest of the build.