Freshen docs for aggregate builders (#3590)
Remove some references to the behavior of older versions of build_runner
and build_resolvers. The changes were made in 2020, and we have had a
major version change since the version which introduced the change.

Change some docs about compatibility with older build_resolvers to
instead describe the utility of splitting steps for improved
invalidation behavior.