Stop depending on analyzer support for Dart <2.12.
It doesn't matter which Dart language version is used by the test
`gives a correct languageVersion based on comments`, since all the
test is doing is confirming that the language version is properly
reflected in the return value of `resolver.libraryFor`. Previously the
test used language version 2.1; it now uses language version 3.1.

This helps unblock removal of legacy (pre-null-safety) support from
the Dart analyzer.