Use curly braces in if statement, in accordance with upcoming lint rule change.
This lint rule currently measures from the end of the condition, but needs to measure from the start of the if-statement. E.g.

if ("long condition and other long condition" != "another long condition" &&
     "more tests".isNotEmpty) return null;
This is currently allowed, but should not be, as per Effective Dart.

This will be fixed in a Dart SDK CL.

PiperOrigin-RevId: 647347218