Fix unintended_html_in_doc_comment lints (#3734)
There are a few that are false positives, collection types with
specified generics in square braces (`[List<int>]`) work in dartdoc, but
it's slightly better style for these cases to not over-link to core
types. Refer to collection types in prose, and use backticks where
describing but not linking to a type is appropriate.

Reformat a code block to use 4 space indent and a blank following the
prose block to use the intended markdown syntax.

Rephrase a few updated docs for readability.