Limit which health checks run
The `breaking` and `leaking` health checks are not currently compatible
with repos using pub workspaces.

Keep `coverage` for now since it's not obvious it's a problem. We may
want to revisit duplicate testing between coverage and the other test
configurations.

Run `version`, `changelog`, and `do-not-submit` checks which are
sensible for every package.