Fix relative imports in wasm loader (#3752)
Currently, the loader script emitted by `build_web_compilers` resolves URLs against the base-URL of the active document. This breaks when the entrypoint script is not in the same directory as the page, e.g. in

```
web/
  index.html
  src/
    app.dart
```

To fix this, this PR changes the logic to resolve URLs against the loader script (which would be `/src/app.dart.js` in the example, properly resolving the other dart2wasm/dart2js components).
I've also added an integration test using a `<script>` tag pointing to a subdirectory to make sure that this is working now.

Closes https://github.com/dart-lang/build/issues/3751