Batch writes (#3771)
* Use batches for builds

* Add changelog entry to build runner

* Review feedback
aAnother attempt for https://github.com/dart-lang/build/issues/3321, my [previous PR](https://github.com/dart-lang/build/pull/3418) for this got stale, this one is a bit simpler.

This attempts to solve the issue of `build_runner` and external tools watching for file system events (like the analysis server) not playing too well together. At the moment, a common issue is that, as a user runs `build_runner build` and the generated assets are written as they are generated, the analysis server will see many file changes and keep re-analyzing sources in an incomplete state.
By making `build_runner` cache pending writes in memory and then flushing them after a completed build, we'll likely get all changes to disk before they're picked up by external tools, reducing friction.

I've implemented this at a fairly low level (the raw file readers and writers making up the default `IOEnvironment`). They have to opt-in to batches by checking for the presence of a zone key when reading or writing assets. We support batches by default when not in low-resources mode. A batch runs in a zone wrapping a whole build run.

A todo is that I need to write integration tests for this, but it would be good to get a review for the overall approach before completing this.

Closes https://github.com/dart-lang/build/issues/3321.
* Bump version of build_runner_core

* Make batch private, fix dependency resolution

* Start fixing tests

* Add lints dependency for out-of-workspace pkg