Consider SmartFake args as always non-matching
If a `SmartFake` instance is passed as an argument to a mock it will be stored
and later may be compared against real instances of the class that was being
faked. A `SmartFake` should never be "used", including being compared to other
objects. Always consider a `SmartFake` actual argument to be non-matching
against any other value arguments. A `SmartFake` may still be a match in some
cases, such as an `any` argument.

An upcoming change in `package:matcher` will make equality operator calls that
throw test errors instead of ignorable non-matches. A `SmartFake` with no fields
that can be read will likely cause an exception when passed to any non-default
equality operator.

PiperOrigin-RevId: 844873060